Add async version of nor flash
Add async version of the nor_flash traits, that can be used in an
async context. In the same way as embedded_hal_async in relation to
embedded_hal, the traits are stored in a separate crate, with the
intention of moving to the main crate once the async GAT feature is part
of Rust stable.